isolate GaugeController from Pools (not deployed on Polygon)
